@@ -4,71 +4,74 @@ import {
   matchTemplate,
 } from "@aws-cdk/assert";
 import { App, DefaultStackSynthesizer } from "@aws-cdk/core";
-import * as fs from "fs";
-import * as path from "path";
+import { readFileSync } from "fs";
+import { resolve } from "path";
 import { BuildConfig } from "../lib/build/buildConfig";
 import { AwsSsoExtensionsForEnterprise } from "../lib/stacks/pipeline/aws-sso-extensions-for-enterprise";
-const yaml = require("js-yaml");
+import yaml = require("js-yaml");
 
 test("Empty Stack", () => {
   const app = new App();
 
   function ensureString(
+    /* eslint-disable  @typescript-eslint/no-explicit-any */
     object: { [name: string]: any },
     propName: string
   ): string {
-    if (!object[propName] || object[propName].trim().length === 0)
+    if (!object[`${propName}`] || object[`${propName}`].trim().length === 0)
       throw new Error(propName + " does not exist or is empty");
 
-    return object[propName];
+    return object[`${propName}`];
   }
 
   function ensureValidString(
+    /* eslint-disable  @typescript-eslint/no-explicit-any */
     object: { [name: string]: any },
     propName: string,
     validList: Array<string>
   ): string {
     if (
-      !object[propName] ||
-      object[propName].trim().length === 0 ||
-      typeof object[propName] !== "string"
+      !object[`${propName}`] ||
+      object[`${propName}`].trim().length === 0 ||
+      typeof object[`${propName}`] !== "string"
     )
       throw new Error(
         propName +
           " does not exist or is empty or is of not the correct data type"
       );
 
-    const value = ("" + object[propName]).toUpperCase();
+    const value = ("" + object[`${propName}`]).toUpperCase();
     if (!validList.includes(value)) {
       throw new Error(
         `${propName} is not one of the valid values - ${validList.toString()}`
       );
     }
 
-    return object[propName];
+    return object[`${propName}`];
   }
 
   function ensureBoolean(
+    /* eslint-disable  @typescript-eslint/no-explicit-any */
     object: { [name: string]: any },
     propName: string
   ): boolean {
-    if (typeof object[propName] !== "boolean")
+    if (typeof object[`${propName}`] !== "boolean")
       throw new Error(
         propName + " does not exist or is of not the correct data type"
       );
 
-    return object[propName];
+    return object[`${propName}`];
   }
 
   function getConfig() {
-    let env = app.node.tryGetContext("config");
+    const env = app.node.tryGetContext("config");
     if (!env)
       throw new Error(
         "Context variable missing on CDK command. Pass in as `-c config=XXX`"
       );
-
-    let unparsedEnv = yaml.load(
-      fs.readFileSync(path.resolve("./config/" + env + ".yaml"), "utf8")
+    /* eslint-disable  @typescript-eslint/no-explicit-any */
+    const unparsedEnv: any = yaml.load(
+      readFileSync(resolve("./config/" + env + ".yaml"), "utf8")
     );
 
     return {
@@ -139,7 +142,6 @@ test("Empty Stack", () => {
           unparsedEnv["Parameters"],
           "PermissionSetCallerRoleArn"
         ),
-        ApiCorsOrigin: ensureString(unparsedEnv["Parameters"], "ApiCorsOrigin"),
         NotificationEmail: ensureString(
           unparsedEnv["Parameters"],
           "NotificationEmail"
@@ -150,7 +152,7 @@ test("Empty Stack", () => {
     };
   }
 
-  let buildConfig: BuildConfig = getConfig();
+  const buildConfig: BuildConfig = getConfig();
 
   // WHEN
   const stack = new AwsSsoExtensionsForEnterprise(
