@@ -8,16 +8,12 @@ gracefully handled by the proxy API
 import {
   AccessLogFormat,
   AuthorizationType,
-  IResource,
   LambdaIntegration,
   LambdaRestApi,
   LogGroupLogDestination,
-  MockIntegration,
-  PassthroughBehavior,
-  ResponseType,
 } from "@aws-cdk/aws-apigateway";
 import { Effect, IRole, PolicyStatement, Role } from "@aws-cdk/aws-iam";
-import { Function } from "@aws-cdk/aws-lambda";
+import { NodejsFunction } from "@aws-cdk/aws-lambda-nodejs";
 import { LogGroup, RetentionDays } from "@aws-cdk/aws-logs";
 import { CfnOutput, Construct } from "@aws-cdk/core";
 import { BuildConfig } from "../build/buildConfig";
@@ -29,7 +25,7 @@ function name(buildConfig: BuildConfig, resourcename: string): string {
 export interface LambdaProxyAPIProps {
   apiNameKey: string;
   apiResourceName: string;
-  proxyfunction: Function;
+  proxyfunction: NodejsFunction;
   apiCallerRoleArn: string;
   methodtype: string;
   apiEndPointReaderAccountID: string;
@@ -52,10 +48,6 @@ export class LambdaProxyAPI extends Construct {
       this,
       name(buildConfig, `${lambdaProxyAPIProps.apiNameKey}-logGroup`),
       {
-        logGroupName: name(
-          buildConfig,
-          `${lambdaProxyAPIProps.apiNameKey}-logGroup`
-        ),
         retention: RetentionDays.ONE_MONTH,
       }
     );
@@ -88,24 +80,9 @@ export class LambdaProxyAPI extends Construct {
       }
     );
 
-    this.lambdaProxyAPI.addGatewayResponse("not_authorized", {
-      type: ResponseType.DEFAULT_4XX,
-      responseHeaders: {
-        "Access-Control-Allow-Origin": "'*'",
-        "Access-Control-Allow-Headers": "'*'",
-        "Access-Control-Allow-Credentials": "'true'",
-      },
-      templates: {
-        "application/json":
-          '{ "message": $context.error.messageString, "type": "$context.error.responseType", "code": "$context.error.responseType" }',
-      },
-    });
-
     const lambdaproxyAPIResource = this.lambdaProxyAPI.root.addResource(
       lambdaProxyAPIProps.apiResourceName
     );
-
-    addCorsOptions(lambdaproxyAPIResource);
 
     this.lambdaProxyAPIRole = Role.fromRoleArn(
       this,
@@ -132,45 +109,5 @@ export class LambdaProxyAPI extends Construct {
         resources: [lambdaProxyAPIMethod.methodArn],
       })
     );
-
-    function addCorsOptions(apiResource: IResource) {
-      apiResource.addMethod(
-        "OPTIONS",
-        new MockIntegration({
-          integrationResponses: [
-            {
-              statusCode: "200",
-              responseParameters: {
-                "method.response.header.Access-Control-Allow-Headers":
-                  "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'",
-                "method.response.header.Access-Control-Allow-Origin":
-                  buildConfig.Parameters.ApiCorsOrigin,
-                "method.response.header.Access-Control-Allow-Credentials":
-                  "'false'",
-                "method.response.header.Access-Control-Allow-Methods":
-                  "'POST,OPTIONS'",
-              },
-            },
-          ],
-          passthroughBehavior: PassthroughBehavior.WHEN_NO_MATCH,
-          requestTemplates: {
-            "application/json": '{"statusCode": 200}',
-          },
-        }),
-        {
-          methodResponses: [
-            {
-              statusCode: "200",
-              responseParameters: {
-                "method.response.header.Access-Control-Allow-Headers": true,
-                "method.response.header.Access-Control-Allow-Methods": true,
-                "method.response.header.Access-Control-Allow-Credentials": true,
-                "method.response.header.Access-Control-Allow-Origin": true,
-              },
-            },
-          ],
-        }
-      );
-    }
   }
 }
