@@ -14,11 +14,12 @@ import {
 } from "@aws-cdk/aws-dynamodb";
 import { Role } from "@aws-cdk/aws-iam";
 import { Key } from "@aws-cdk/aws-kms";
-import * as lambda from "@aws-cdk/aws-lambda"; //Needed to avoid semgrep throwing up https://cwe.mitre.org/data/definitions/95.html
+import { LayerVersion, Runtime } from "@aws-cdk/aws-lambda";
+import { NodejsFunction } from "@aws-cdk/aws-lambda-nodejs";
 import { Bucket, EventType } from "@aws-cdk/aws-s3";
 import { LambdaDestination } from "@aws-cdk/aws-s3-notifications";
 import { CfnOutput, Construct, RemovalPolicy } from "@aws-cdk/core";
-import * as Path from "path";
+import { join } from "path";
 import { BuildConfig } from "../build/buildConfig";
 import { LambdaProxyAPI } from "./lambda-proxy-api";
 
@@ -27,7 +28,7 @@ function name(buildConfig: BuildConfig, resourcename: string): string {
 }
 
 export interface PermissionSetCRUDProps {
-  readonly nodeJsLayer: lambda.LayerVersion;
+  readonly nodeJsLayer: LayerVersion;
   readonly linksTableName: string;
   readonly errorNotificationsTopicArn: string;
   readonly ssoArtefactsBucket: Bucket;
@@ -38,10 +39,10 @@ export interface PermissionSetCRUDProps {
 export class PermissionSetCRUD extends Construct {
   public readonly permissionSetTable: Table;
   public readonly permissionSetArnTable: Table;
-  public readonly permissionSetAPIHandler: lambda.Function;
+  public readonly permissionSetAPIHandler: NodejsFunction;
   public readonly permissionSetAPI: LambdaRestApi;
-  public readonly permissionSetCuHandler: lambda.Function;
-  public readonly permissionSetDelHandler: lambda.Function;
+  public readonly permissionSetCuHandler: NodejsFunction;
+  public readonly permissionSetDelHandler: NodejsFunction;
 
   constructor(
     scope: Construct,
@@ -90,27 +91,33 @@ export class PermissionSetCRUD extends Construct {
       buildConfig.Parameters.PermissionSetProvisioningMode.toLowerCase() ===
       "api"
     ) {
-      this.permissionSetAPIHandler = new lambda.Function(
+      this.permissionSetAPIHandler = new NodejsFunction(
         this,
         name(buildConfig, "psApiHandler"),
         {
-          runtime: lambda.Runtime.NODEJS_14_X,
-          handler: "permissionSetApi.handler",
           functionName: name(buildConfig, "psApiHandler"),
-          code: lambda.Code.fromAsset(
-            Path.join(
-              __dirname,
-              "../",
-              "lambda",
-              "functions",
-              "ddb-import-handlers"
-            )
+          runtime: Runtime.NODEJS_14_X,
+          entry: join(
+            __dirname,
+            "../",
+            "lambda-functions",
+            "ddb-import-handlers",
+            "src",
+            "permissionSetApi.ts"
           ),
+          bundling: {
+            minify: true,
+            externalModules: [
+              "@aws-sdk/client-dynamodb",
+              "@aws-sdk/client-s3",
+              "@aws-sdk/lib-dynamodb",
+              "ajv",
+            ],
+          },
           layers: [PermissionSetCRUDProps.nodeJsLayer],
           environment: {
             DdbTable: this.permissionSetTable.tableName,
             AWS_NODEJS_CONNECTION_REUSE_ENABLED: "1",
-            corsOrigin: buildConfig.Parameters.ApiCorsOrigin,
             linksTable: PermissionSetCRUDProps.linksTableName,
             artefactsBucketName:
               PermissionSetCRUDProps.ssoArtefactsBucket.bucketName,
@@ -133,22 +140,30 @@ export class PermissionSetCRUD extends Construct {
         }
       ).lambdaProxyAPI;
     } else {
-      this.permissionSetCuHandler = new lambda.Function(
+      this.permissionSetCuHandler = new NodejsFunction(
         this,
         name(buildConfig, "psCuHandler"),
         {
-          runtime: lambda.Runtime.NODEJS_14_X,
-          handler: "permissionSetCu.handler",
           functionName: name(buildConfig, "psCuHandler"),
-          code: lambda.Code.fromAsset(
-            Path.join(
-              __dirname,
-              "../",
-              "lambda",
-              "functions",
-              "ddb-import-handlers"
-            )
+          runtime: Runtime.NODEJS_14_X,
+          entry: join(
+            __dirname,
+            "../",
+            "lambda-functions",
+            "ddb-import-handlers",
+            "src",
+            "permissionSetCu.ts"
           ),
+          bundling: {
+            minify: true,
+            externalModules: [
+              "@aws-sdk/client-dynamodb",
+              "@aws-sdk/client-s3",
+              "@aws-sdk/client-sns",
+              "@aws-sdk/lib-dynamodb",
+              "ajv",
+            ],
+          },
           layers: [PermissionSetCRUDProps.nodeJsLayer],
           environment: {
             DdbTable: this.permissionSetTable.tableName,
@@ -182,22 +197,29 @@ export class PermissionSetCRUD extends Construct {
         permissionSetCallerRole
       );
 
-      this.permissionSetDelHandler = new lambda.Function(
+      this.permissionSetDelHandler = new NodejsFunction(
         this,
         name(buildConfig, "psDelHandler"),
         {
-          runtime: lambda.Runtime.NODEJS_14_X,
-          handler: "permissionSetDel.handler",
           functionName: name(buildConfig, "psDelHandler"),
-          code: lambda.Code.fromAsset(
-            Path.join(
-              __dirname,
-              "../",
-              "lambda",
-              "functions",
-              "ddb-import-handlers"
-            )
+          runtime: Runtime.NODEJS_14_X,
+          entry: join(
+            __dirname,
+            "../",
+            "lambda-functions",
+            "ddb-import-handlers",
+            "src",
+            "permissionSetDel.ts"
           ),
+          bundling: {
+            minify: true,
+            externalModules: [
+              "@aws-sdk/client-dynamodb",
+              "@aws-sdk/client-sns",
+              "@aws-sdk/lib-dynamodb",
+              "ajv",
+            ],
+          },
           layers: [PermissionSetCRUDProps.nodeJsLayer],
           environment: {
             DdbTable: this.permissionSetTable.tableName,
